html/semantics/document-metadata/the-meta-element/pragma-directives/http-equiv-enumerated-ascii-case-insensitive.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="help" href="https://html.spec.whatwg.org/#attr-meta-http-equiv">
<link rel="help" href="https://html.spec.whatwg.org/#enumerated-attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function() {
  let loaded = 0;

  // we use a message rather than the iframe’s load event to avoid dealing with
  // spurious load events that some browsers dispatch on the initial about:blank
  addEventListener("message", this.step_func(event => {
    if (++loaded == 3) {
      const iframe = document.querySelectorAll("iframe");

      assert_false(iframe[0].contentWindow.inline);
      assert_false(iframe[1].contentWindow.inline);
      assert_true(iframe[2].contentWindow.inline);

      this.done();
    }
  }));
});
</script>
<iframe src="http-equiv-enumerated-ascii-case-insensitive-a.html"></iframe>
<iframe src="http-equiv-enumerated-ascii-case-insensitive-b.html"></iframe>
<iframe src="http-equiv-enumerated-ascii-case-insensitive-c.html"></iframe>
